Make each test runnable on in isolation
