+++ package.json
@@ -4,7 +4,7 @@
   "description": "REST storage and service dispatcher",
   "main": "server.js",
   "scripts": {
-    "test": "sh test/utils/cleandb.sh && mocha",
+    "test": "sh test/utils/cleandb.sh && mocha --recursive",
     "coverage": "sh test/utils/cleandb.sh && istanbul cover _mocha -- -R spec",
     "coveralls": "cat ./coverage/lcov.info | coveralls"
   },
